refact : custom field queries
