+++ graphql/schemas/Ecosystem/customField.graphql
@@ -19,7 +19,7 @@ extend type Mutation @guard {
         )
     getCustomField(input: CustomFieldInput!): Mixed
         @field(
-            resolver: "App\\GraphQL\\Ecosystem\\Mutations\\CustomFields\\CustomFieldMutation@get"
+            resolver: "App\\GraphQL\\Ecosystem\\Mutations\\CustomFields\\[j]@get"
         )
     getAllCustomField(input: CustomFieldInput!): Mixed
         @field(
@@ -28,5 +28,25 @@ extend type Mutation @guard {
     deleteCustomField(input: CustomFieldInput!): Boolean!
         @field(
             resolver: "App\\GraphQL\\Ecosystem\\Mutations\\CustomFields\\CustomFieldMutation@delete"
+        )
+}
+
+extend type Query @guard {
+    getCustomField(
+        name: String!
+        system_module_uuid: String!
+        entity_id: String!
+    ): Boolean
+        @field(
+            resolver: "App\\GraphQL\\Ecosystem\\Queries\\CustomFields\\CustomFieldQuery@get"
+        )
+    getAllCustomField(
+        name: String!
+        system_module_uuid: String!
+        entity_id: String!
+    ): Boolean
+        @field(
+            resolver:
+              "App\\GraphQL\\Ecosystem\\Queries\\CustomFields\\CustomFieldQuery@getAll"
         )
 }
