+++ graphql/schemas/Ecosystem/visitor.graphql
